Fix literal type in bitwise left shift operations (#67)
This commit casts the occurences of literal `1` where we bitwise left
shift it to the appropriate type (unsigned long, size_t, uintptr_t).

With this change we avoid erroneous results when running above 32 no-
des. The maximum number of nodes ArgoDSM can run at the moment is 64.

+++ src/backend/mpi/coherence.cpp
@@ -82,7 +82,7 @@ namespace argo {
 			std::size_t argo_address =
 				((reinterpret_cast<std::size_t>(addr)-start_address)/block_size)*block_size;
 			const node_id_t node_id = argo::backend::node_id();
-			const std::size_t node_id_bit = 1 << node_id;
+			const std::size_t node_id_bit = static_cast<std::size_t>(1) << node_id;
 
 			// Lock relevant mutexes. Start statistics timekeeping
 			double t1 = MPI_Wtime();
